% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_agency_contracts}
\alias{wag_agency_contracts}
\title{State agency contracts (current fiscal year)}
\usage{
wag_agency_contracts(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
State agency contracts (current fiscal year)
}
