% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_apprentice_utilization}
\alias{wag_apprentice_utilization}
\title{L&I apprentice utilization on public works}
\usage{
wag_apprentice_utilization(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
L&I apprentice utilization on public works
}
