% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_campaign_summary}
\alias{wag_campaign_summary}
\title{Campaign finance summary}
\usage{
wag_campaign_summary(filer_name = NULL, year = NULL, limit = 1000)
}
\arguments{
\item{filer_name}{Filter by name}

\item{year}{Election year}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Campaign finance summary
}
