% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_cannabis_renewals}
\alias{wag_cannabis_renewals}
\title{LCB cannabis license renewals}
\usage{
wag_cannabis_renewals(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
LCB cannabis license renewals
}
