% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_consumer_complaints}
\alias{wag_consumer_complaints}
\title{Consumer complaints to Attorney General}
\usage{
wag_consumer_complaints(business_name = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{business_name}{Filter by business}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Consumer complaints to Attorney General
}
