% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_contract_sales}
\alias{wag_contract_sales}
\title{Statewide master contract sales data}
\usage{
wag_contract_sales(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Statewide master contract sales data
}
