% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_contractor_licenses}
\alias{wag_contractor_licenses}
\title{L&I contractor license data}
\usage{
wag_contractor_licenses(business_name = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{business_name}{Filter by business name}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
L&I contractor license data
}
