% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_contributions}
\alias{wag_contributions}
\title{Campaign contributions to candidates and political committees}
\usage{
wag_contributions(filer_name = NULL, year = NULL,
                              where = NULL, limit = 1000)
}
\arguments{
\item{filer_name}{Filter by candidate/committee name (SoQL LIKE)}

\item{year}{Election year filter}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Washington Public Disclosure Commission (PDC) data on contributions.
}
