% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_cpa_firms}
\alias{wag_cpa_firms}
\title{Washington State Certified Public Accountants}
\usage{
wag_cpa_firms(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Washington State Certified Public Accountants
}
