% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_crime_reports}
\alias{wag_crime_reports}
\title{Criminal justice data book}
\usage{
wag_crime_reports(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Criminal justice data book
}
