% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_data_breaches}
\alias{wag_data_breaches}
\title{Data breach notifications affecting WA residents}
\usage{
wag_data_breaches(year = NULL, industry_type = NULL,
                              where = NULL, limit = 1000)
}
\arguments{
\item{year}{Filter by year}

\item{industry_type}{Filter by industry}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Data breach notifications affecting WA residents
}
