% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_enforcement_cases}
\alias{wag_enforcement_cases}
\title{PDC enforcement cases}
\usage{
wag_enforcement_cases(case_status = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{case_status}{Filter: "Open", "Closed", etc.}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
PDC enforcement cases
}
