% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_expenditures}
\alias{wag_expenditures}
\title{Campaign expenditures by candidates and political committees}
\usage{
wag_expenditures(filer_name = NULL, year = NULL,
                             where = NULL, limit = 1000)
}
\arguments{
\item{filer_name}{Filter by candidate/committee name}

\item{year}{Election year}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Campaign expenditures by candidates and political committees
}
