% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_fish_plants}
\alias{wag_fish_plants}
\title{WDFW fish releases/plants}
\usage{
wag_fish_plants(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
WDFW fish releases/plants
}
