% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_graduation_rates}
\alias{wag_graduation_rates}
\title{Graduation pathway completion rates}
\usage{
wag_graduation_rates(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Graduation pathway completion rates
}
