% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_greenhouse_gas}
\alias{wag_greenhouse_gas}
\title{Greenhouse gas emissions facility data (2017)}
\usage{
wag_greenhouse_gas(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Greenhouse gas emissions facility data (2017)
}
