% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_hatcheries}
\alias{wag_hatcheries}
\title{WDFW hatchery information}
\usage{
wag_hatcheries(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
WDFW hatchery information
}
