% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_it_spend}
\alias{wag_it_spend}
\title{IT spend by technology towers}
\usage{
wag_it_spend(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
IT spend by technology towers
}
