% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_lobbyist_compensation}
\alias{wag_lobbyist_compensation}
\title{Lobbyist compensation and expenses by source}
\usage{
wag_lobbyist_compensation(year = NULL, limit = 1000)
}
\arguments{
\item{year}{Filter year}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Lobbyist compensation and expenses by source
}
