% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_lobbyist_registrations}
\alias{wag_lobbyist_registrations}
\title{Lobbyist employment registrations}
\usage{
wag_lobbyist_registrations(year = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{year}{Registration year}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Lobbyist employment registrations
}
