% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_officer_certifications}
\alias{wag_officer_certifications}
\title{Criminal justice training commission officer certification cases}
\usage{
wag_officer_certifications(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Criminal justice training commission officer certification cases
}
