% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_professional_licenses}
\alias{wag_professional_licenses}
\title{Professional license transactions (DOL)}
\usage{
wag_professional_licenses(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Professional license transactions (DOL)
}
