% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_public_works_intents}
\alias{wag_public_works_intents}
\title{L&I public works intent projects}
\usage{
wag_public_works_intents(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
L&I public works intent projects
}
