% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_salmon_populations}
\alias{wag_salmon_populations}
\title{WDFW salmonid population indicators}
\usage{
wag_salmon_populations(species = NULL, where = NULL, limit = 1000)
}
\arguments{
\item{species}{Filter by species name}

\item{where}{Additional SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
WDFW salmonid population indicators
}
