% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_school_improvement}
\alias{wag_school_improvement}
\title{Washington School Improvement Framework (WSIF)}
\usage{
wag_school_improvement(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Washington School Improvement Framework (WSIF)
}
