% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_sepa}
\alias{wag_sepa}
\title{SEPA (State Environmental Policy Act) register}
\usage{
wag_sepa(where = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{q}{Full-text search}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
SEPA (State Environmental Policy Act) register
}
