% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_spawning_surveys}
\alias{wag_spawning_surveys}
\title{WDFW spawning ground surveys}
\usage{
wag_spawning_surveys(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
WDFW spawning ground surveys
}
