% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa.R
\name{wag_vehicle_registrations}
\alias{wag_vehicle_registrations}
\title{Vehicle registrations by class and county}
\usage{
wag_vehicle_registrations(where = NULL, limit = 1000)
}
\arguments{
\item{where}{SoQL where clause}

\item{limit}{Max rows (default 1000)}

}
\value{
tibble
}
\description{
Vehicle registrations by class and county
}
